Restructured extract transformation (and some related classes).
