+++ src/psyclone/tests/gocean0p1_transformations_test.py
@@ -14,9 +14,8 @@ import pytest
 from psyclone.configuration import Config
 from psyclone.parse.algorithm import parse
 from psyclone.psyGen import PSyFactory
-from psyclone.transformations import TransformationError,\
-    LoopFuseTrans,\
-    GOceanLoopFuseTrans,\
+from psyclone.psyir.transformations import TransformationError
+from psyclone.transformations import LoopFuseTrans, GOceanLoopFuseTrans, \
     GOceanOMPParallelLoopTrans
 
 API = "gocean0.1"
