+++ src/psyclone/tests/profile_test.py
@@ -45,8 +45,9 @@ import pytest
 from psyclone.generator import GenerationError
 from psyclone.profiler import Profiler, ProfileNode
 from psyclone.psyGen import Loop, NameSpace, InternalError
+from psyclone.psyir.transformations import TransformationError
 from psyclone.transformations import GOceanOMPLoopTrans, OMPParallelTrans, \
-    ProfileRegionTrans, TransformationError
+    ProfileRegionTrans
 from psyclone.tests.utilities import get_invoke
