+++ src/psyclone/transformations.py
@@ -48,6 +48,7 @@ from psyclone.configuration import Config
 from psyclone.undoredo import Memento
 from psyclone.dynamo0p3 import VALID_ANY_SPACE_NAMES, \
     VALID_ANY_DISCONTINUOUS_SPACE_NAMES
+from psyclone.psyir.transformations import RegionTrans, TransformationError
 
 VALID_OMP_SCHEDULES = ["runtime", "static", "dynamic", "guided", "auto"]
 
@@ -78,144 +79,6 @@ def check_intergrid(node):
                 "This Transformation cannot currently be applied to nodes "
                 "which have inter-grid kernels as descendents and {0} is "
                 "such a kernel.".format(kern.name))
-
-
-class TransformationError(Exception):
-    ''' Provides a PSyclone-specific error class for errors found during
-        code transformation operations. '''
-
-    def __init__(self, value):
-        Exception.__init__(self, value)
-        self.value = "Transformation Error: "+value
-
-    def __str__(self):
-        return repr(self.value)
-
-
-@six.add_metaclass(abc.ABCMeta)
-class RegionTrans(Transformation):
-    '''
-    This abstract class is a base class for all transformations that act
-    on a list of nodes. It gives access to a _validate function that
-    makes sure that the nodes in the list are in the same order as in
-    the original AST, no node is duplicated, and that all nodes have
-    the same parent. We also check that all nodes to be enclosed are
-    valid for this transformation - this requires that the sub-class
-    populate the `valid_node_types` tuple.
-
-    '''
-    # The types of Node that we support within this region. Must be
-    # populated by sub-class.
-    valid_node_types = ()
-
-    # Avoid pylint warning about abstract functions (apply, name) not
-    # overwritten:
-    # pylint: disable=abstract-method,arguments-differ
-
-    def validate(self, node_list, options=None):
-        '''
-        Checks that the nodes in node_list are valid for a region
-        transformation.
-
-        :param node_list: list of PSyIR nodes or a single Schedule.
-        :type node_list: :py:class:`psyclone.psyGen.Schedule` or a \
-                         list of :py:class:`psyclone.psyGen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-        :param bool options["node-type-check"]: this flag controls if the \
-                type of the nodes enclosed in the region should be tested \
-                to avoid using unsupported nodes inside a region.
-
-        :raises TransformationError: if the nodes in the list are not \
-                in the original order in which they are in the AST, \
-                a node is duplicated or the nodes have different parents.
-        :raises TransformationError: if any of the nodes to be enclosed in \
-                the region are of an unsupported type.
-        :raises TransformationError: if the parent of the supplied Nodes is \
-                                     not a Schedule or a Directive.
-        :raises TransformationError: if the nodes are in a NEMO Schedule and \
-                                     the transformation acts on the child of \
-                                     a single-line If or Where statment.
-
-        '''
-        # pylint: disable=too-many-branches
-        from psyclone.psyGen import IfBlock, Loop
-        from psyclone.nemo import NemoInvokeSchedule
-        if not options:
-            options = {}
-        node_parent = node_list[0].parent
-        prev_position = -1
-        for child in node_list:
-            if child.parent is not node_parent:
-                raise TransformationError(
-                    "Error in {0} transformation: supplied nodes "
-                    "are not children of the same parent."
-                    .format(self.name))
-            if prev_position >= 0 and prev_position+1 != child.position:
-                raise TransformationError(
-                    "Children are not consecutive children of one parent: "
-                    "child '{0}' has position {1}, but previous child had "
-                    "position {2}."
-                    .format(str(child), child.position, prev_position))
-            prev_position = child.position
-
-        # Check that the proposed region contains only supported node types
-        if options.get("node-type-check", True):
-            for child in node_list:
-                # Stop at any instance of Kern to avoid going into the
-                # actual kernels, e.g. in Nemo inlined kernels
-                flat_list = [item for item in child.walk(object, Kern)
-                             if not isinstance(item, Schedule)]
-                for item in flat_list:
-                    if not isinstance(item, self.valid_node_types):
-                        raise TransformationError(
-                            "Nodes of type '{0}' cannot be enclosed by a {1} "
-                            "transformation".format(type(item), self.name))
-
-        # If we've been passed a list that contains one or more Schedules
-        # then something is wrong. e.g. two Schedules that are both children
-        # of an IfBlock would imply that the transformation is being applied
-        # around both the if-body and the else-body and that doesn't make
-        # sense.
-        if isinstance(node_list, list) and len(node_list) > 1 and \
-           any([isinstance(node, Schedule) for node in node_list]):
-            raise TransformationError(
-                "Cannot apply a transformation to multiple nodes when one or "
-                "more is a Schedule. Either target a single Schedule or the"
-                " children of a Schedule.")
-
-        # Sanity check that we've not been passed the condition part of
-        # an If statement or the bounds of a Loop. If the parent node is
-        # a Loop of IfBlock then we can only accept a single Schedule.
-        # TODO #542 Once everything has a Schedule we can tidy this up
-        # a little by requiring that either the parent be a Schedule or
-        # that the node-list consists of a single Schedule.
-        if isinstance(node_parent, (Loop, IfBlock)) and \
-           not isinstance(node_list[0], Schedule):
-            # We've already checked for lists with len > 1 that contain a
-            # Schedule above so if the first item is a Schedule then that's
-            # all the list contains.
-            raise TransformationError(
-                "Cannot apply transformation to the immediate children of a "
-                "Loop/IfBlock unless it is to a single Schedule representing"
-                " the Loop/If/Else body.")
-
-        # The checks below this point only apply to the NEMO API and can be
-        # removed once #435 is done.
-        node = node_list[0]
-        if not isinstance(node.root, NemoInvokeSchedule):
-            return
-
-        if_or_loop = node.ancestor((IfBlock, Loop))
-        if if_or_loop and ("was_single_stmt" in if_or_loop.annotations
-                           or "was_where" in if_or_loop.annotations):
-            # This limitation is because the NEMO API currently relies on
-            # manipulation of the fparser2 parse tree
-            # TODO #435.
-            raise TransformationError(
-                "In the NEMO API a transformation cannot be applied to the "
-                "children of either a single-line if statement or a PSyIR loop"
-                " representing a WHERE construct.")
 
 
 class KernelTrans(Transformation):
@@ -4077,295 +3940,3 @@ class NemoExplicitLoopTrans(Transformation):
             raise TransformationError(
                 "Cannot apply NemoExplicitLoopTrans to something that is "
                 "not a NemoImplicitLoop (got {0})".format(type(loop)))
-
-
-class ExtractRegionTrans(RegionTrans):
-    ''' Provides a transformation to extract code represented by a \
-    subset of the Nodes in the PSyIR of a Schedule into a stand-alone \
-    program. Examples are given in descriptions of children classes \
-    DynamoExtractRegionTrans and GOceanExtractRegionTrans.
-
-    After applying the transformation the Nodes marked for extraction are \
-    children of the ExtractNode. \
-    Nodes to extract can be individual constructs within an Invoke (e.g. \
-    Loops containing a Kernel or BuiltIn call) or entire Invokes. This \
-    functionality does not support distributed memory.
-    '''
-    from psyclone import psyGen
-    # The types of node that this transformation can enclose
-    valid_node_types = (psyGen.Loop, psyGen.Kern, psyGen.BuiltIn,
-                        psyGen.Directive, psyGen.Literal, psyGen.Reference)
-
-    def __str__(self):
-        return ("Create a sub-tree of the PSyIR that has ExtractNode "
-                "at its root.")
-
-    @property
-    def name(self):
-        ''' Returns the name of this transformation as a string.'''
-        return "ExtractRegionTrans"
-
-    def validate(self, node_list, options=None):
-        ''' Perform validation checks before applying the transformation
-
-        :param node_list: the list of Node(s) we are checking.
-        :type node_list: list of :py:class:`psyclone.psyGen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if distributed memory is configured.
-        :raises TransformationError: if transformation is applied to a \
-                                     Kernel or a BuiltIn call without its \
-                                     parent Loop.
-        :raises TransformationError: if transformation is applied to a Loop \
-                                     without its parent Directive when \
-                                     optimisations are applied.
-        :raises TransformationError: if transformation is applied to an \
-                                     orphaned Directive without its parent \
-                                     Directive.
-        '''
-
-        # First check constraints on Nodes in the node_list common to
-        # all RegionTrans transformations.
-        super(ExtractRegionTrans, self).validate(node_list, options)
-
-        # Now check ExtractRegionTrans specific constraints.
-
-        # Extracting distributed memory code is not supported due to
-        # generation of infrastructure calls to set halos dirty or clean.
-        # This constraint covers the presence of HaloExchange and
-        # GlobalSum classses as they are only generated when distributed
-        # memory is enabled.
-        if Config.get().distributed_memory:
-            raise TransformationError(
-                "Error in {0}: Distributed memory is not supported."
-                .format(str(self.name)))
-
-        # Check constraints not covered by valid_node_types for
-        # individual Nodes in node_list.
-        from psyclone.psyGen import Loop, BuiltIn, Directive, \
-            OMPParallelDirective, ACCParallelDirective
-
-        for node in node_list:
-
-            # Check that ExtractNode is not inserted between a Kernel or
-            # a BuiltIn call and its parent Loop.
-            if isinstance(node, (Kern, BuiltIn)) and \
-               isinstance(node.parent.parent, Loop):
-                raise TransformationError(
-                    "Error in {0}: Extraction of a Kernel or a Built-in "
-                    "call without its parent Loop is not allowed."
-                    .format(str(self.name)))
-
-            # Check that ExtractNode is not inserted between a Loop and its
-            # parent Directive when optimisations are applied, as this may
-            # result in including the end Directive for extraction but
-            # not the beginning.
-            if isinstance(node, Loop) and isinstance(node.parent, Schedule) \
-               and isinstance(node.parent.parent, Directive):
-                raise TransformationError(
-                    "Error in {0}: Extraction of a Loop without its parent "
-                    "Directive is not allowed.".format(str(self.name)))
-
-            # Check that ExtractNode is not inserted within a thread
-            # parallel region when optimisations are applied. For instance,
-            # this may be between an orphaned Directive (e.g. OMPDoDirective,
-            # ACCLoopDirective) and its ancestor Directive (e.g. ACC or OMP
-            # Parallel Directive) or within an OMPParallelDoDirective.
-            if node.ancestor(OMPParallelDirective) or \
-                    node.ancestor(ACCParallelDirective):
-                raise TransformationError(
-                    "Error in {0}: Extraction of Nodes enclosed within "
-                    "a thread parallel region is not allowed."
-                    .format(str(self.name)))
-
-    def apply(self, nodes, options=None):
-        # pylint: disable=arguments-differ
-        ''' Apply this transformation to a subset of the Nodes within
-        a Schedule - i.e. enclose the specified Nodes in the Schedule
-        within a single Extract region.
-
-        :param nodes: a single Node or a list of Nodes.
-        :type nodes: (list of) :py:class:`psyclone.psyGen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if the `nodes` argument is not of \
-                                     the correct type.
-
-        :returns: tuple of the modified Schedule and a record of the \
-                  transformation.
-        :rtype: (:py:class:`psyclone.psyGen.Schedule`, \
-                 :py:class:`psyclone.undoredo.Memento`).
-        '''
-
-        # Check whether we've been passed a list of Nodes or just a
-        # single Node. If the latter then we create ourselves a list
-        # containing just that Node.
-        from psyclone.psyGen import Node
-        if isinstance(nodes, list) and isinstance(nodes[0], Node):
-            node_list = nodes
-        elif isinstance(nodes, Schedule):
-            node_list = nodes.children
-        elif isinstance(nodes, Node):
-            node_list = [nodes]
-        else:
-            arg_type = str(type(nodes))
-            raise TransformationError("Error in {0}: "
-                                      "Argument must be a single Node in a "
-                                      "Schedule or a list of Nodes in a "
-                                      "Schedule but have been passed an "
-                                      "object of type: {1}".
-                                      format(str(self.name), arg_type))
-
-        # Validate transformation
-        self.validate(node_list, options)
-
-        # Keep a reference to the parent of the Nodes that are to be
-        # enclosed within an Extract region. Also keep the index of
-        # the first child to be enclosed as that will be the position
-        # of the ExtractNode.
-        node_parent = node_list[0].parent
-        node_position = node_list[0].position
-
-        # Create a Memento of the Schedule and the proposed
-        # transformation
-        schedule = node_list[0].root
-
-        keep = Memento(schedule, self)
-
-        from psyclone.extractor import ExtractNode
-        extract_node = ExtractNode(parent=node_parent, children=node_list[:])
-
-        # Change all of the affected children so that they have the
-        # ExtractNode as their parent. Use a slice of the list of Nodes
-        # so that we're looping over a local copy of the list. Otherwise
-        # things get confused when we remove children from the list.
-        for child in node_list[:]:
-            # Remove child from the parent's list of children
-            node_parent.children.remove(child)
-            child.parent = extract_node
-
-        # Add the ExtractNode as a child of the parent of the Nodes being
-        # enclosed at the original location of the first of these Nodes
-        node_parent.addchild(extract_node,
-                             index=node_position)
-
-        return schedule, keep
-
-
-class DynamoExtractRegionTrans(ExtractRegionTrans):
-    ''' Dynamo0.3 API application of ExtractRegionTrans transformation \
-    to extract code into a stand-alone program. For example:
-
-    >>> from psyclone.parse.algorithm import parse
-    >>> from psyclone.psyGen import PSyFactory
-    >>>
-    >>> API = "dynamo0.3"
-    >>> FILENAME = "solver_alg.x90"
-    >>> ast, invokeInfo = parse(FILENAME, api=API)
-    >>> psy = PSyFactory(API, distributed_memory=False).create(invoke_info)
-    >>> schedule = psy.invokes.get('invoke_0').schedule
-    >>>
-    >>> from psyclone.transformations import DynamoExtractRegionTrans
-    >>> etrans =  DynamoExtractRegionTrans()
-    >>>
-    >>> # Apply DynamoExtractRegionTrans transformation to selected Nodes
-    >>> newsched, _ = etrans.apply(schedule.children[0:3])
-    >>> newsched.view()
-    '''
-
-    @property
-    def name(self):
-        ''' Returns the name of this transformation as a string.'''
-        return "DynamoExtractRegionTrans"
-
-    def validate(self, node_list, options=None):
-        ''' Perform Dynamo0.3 API specific validation checks before applying
-        the transformation.
-
-        :param node_list: the list of Node(s) we are checking.
-        :type node_list: list of :py:class:`psyclone.psyGen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if transformation is applied to a Loop \
-                                     over cells in a colour without its \
-                                     parent Loop over colours.
-        '''
-
-        # First check constraints on Nodes in the node_list inherited from
-        # the parent classes (ExtractRegionTrans and RegionTrans)
-        super(DynamoExtractRegionTrans, self).validate(node_list, options)
-
-        # Check DynamoExtractRegionTrans specific constraints
-        from psyclone.dynamo0p3 import DynLoop
-        for node in node_list:
-
-            # Check that ExtractNode is not inserted between a Loop
-            # over colours and a Loop over cells in a colour when
-            # colouring is applied.
-            ancestor = node.ancestor(DynLoop)
-            if ancestor and ancestor.loop_type == 'colours':
-                raise TransformationError(
-                    "Error in {0} for Dynamo0.3 API: Extraction of a Loop "
-                    "over cells in a colour without its ancestor Loop over "
-                    "colours is not allowed.".format(str(self.name)))
-
-
-class GOceanExtractRegionTrans(ExtractRegionTrans):
-    ''' GOcean1.0 API application of ExtractRegionTrans transformation \
-    to extract code into a stand-alone program. For example:
-
-    >>> from psyclone.parse.algorithm import parse
-    >>> from psyclone.psyGen import PSyFactory
-    >>>
-    >>> API = "gocean1.0"
-    >>> FILENAME = "shallow_alg.f90"
-    >>> ast, invokeInfo = parse(FILENAME, api=API)
-    >>> psy = PSyFactory(API, distributed_memory=False).create(invoke_info)
-    >>> schedule = psy.invokes.get('invoke_0').schedule
-    >>>
-    >>> from psyclone.transformations import GOceanExtractRegionTrans
-    >>> etrans = GOceanExtractRegionTrans()
-    >>>
-    >>> # Apply GOceanExtractRegionTrans transformation to selected Nodes
-    >>> newsched, _ = etrans.apply(schedule.children[0])
-    >>> newsched.view()
-    '''
-
-    @property
-    def name(self):
-        ''' Returns the name of this transformation as a string.'''
-        return "GOceanExtractRegionTrans"
-
-    def validate(self, node_list, options=None):
-        ''' Perform GOcean1.0 API specific validation checks before applying
-        the transformation.
-
-        :param node_list: the list of Node(s) we are checking.
-        :type node_list: list of :py:class:`psyclone.psyGen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if transformation is applied to an \
-                                     inner Loop without its parent outer \
-                                     Loop.
-        '''
-
-        # First check constraints on Nodes in the node_list inherited from
-        # the parent classes (ExtractRegionTrans and RegionTrans)
-        super(GOceanExtractRegionTrans, self).validate(node_list, options)
-
-        # Check GOceanExtractRegionTrans specific constraints
-        from psyclone.gocean1p0 import GOLoop
-        for node in node_list:
-
-            # Check that ExtractNode is not inserted between an inner
-            # and an outer Loop.
-            ancestor = node.ancestor(GOLoop)
-            if ancestor and ancestor.loop_type == 'outer':
-                raise TransformationError(
-                    "Error in {0} for GOcean1.0 API: Extraction of an "
-                    "inner Loop without its ancestor outer Loop is not "
-                    "allowed.".format(str(self.name)))
